refactor: extract renderLabels pure js logic into function
