+++ src/Chart/renderLabels.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import DefaultLabel from '../Label';
+import Label from '../Label';
 import {
   bisectorAngle,
   extractPercentage,
@@ -8,11 +8,7 @@ import {
 } from '../utils';
 import type { PropsWithDefaults as ChartProps } from './Chart';
 import type { LabelRenderProps } from '../Label';
-import type {
-  ExtendedData,
-  LabelRenderFunction,
-  BaseDataEntry,
-} from '../commonTypes';
+import type { ExtendedData, BaseDataEntry } from '../commonTypes';
 
 function round(number: number): number {
   const divisor = 1e14; // 14 decimals
@@ -46,32 +42,10 @@ function evaluateTextAnchorPosition({
   return 'middle';
 }
 
-function renderLabelElement<DataEntry extends BaseDataEntry>(
-  renderLabel: LabelRenderFunction<DataEntry>,
-  labelProps: LabelRenderProps<DataEntry>
-): JSX.Element | null {
-  const label = renderLabel(labelProps);
-  if (typeof label === 'string' || typeof label === 'number') {
-    return (
-      <DefaultLabel
-        key={`label-${labelProps.dataEntry.key || labelProps.dataIndex}`}
-        {...labelProps}
-      >
-        {label}
-      </DefaultLabel>
-    );
-  }
-
-  if (React.isValidElement(label)) {
-    return label;
-  }
-  return null;
-}
-
-export default function renderLabels<DataEntry extends BaseDataEntry>(
+function makeLabelRenderProps<DataEntry extends BaseDataEntry>(
   data: ExtendedData<DataEntry>,
   props: ChartProps<DataEntry>
-) {
+): LabelRenderProps<DataEntry>[] {
   return data.map((dataEntry, index) => {
     const segmentsShift = functionProp(props.segmentsShift, index) ?? 0;
     const distanceFromCenter =
@@ -82,7 +56,7 @@ export default function renderLabels<DataEntry extends BaseDataEntry>(
     );
 
     // This object is passed as argument to the "label" function prop
-    const labelRenderProps = {
+    const labelRenderProps: LabelRenderProps<DataEntry> = {
       x: props.center[0],
       y: props.center[1],
       dx,
@@ -97,6 +71,18 @@ export default function renderLabels<DataEntry extends BaseDataEntry>(
       style: functionProp(props.labelStyle, index),
     };
 
-    return props.label && renderLabelElement(props.label, labelRenderProps);
+    return labelRenderProps;
   });
+}
+
+export default function renderLabels<DataEntry extends BaseDataEntry>(
+  data: ExtendedData<DataEntry>,
+  props: ChartProps<DataEntry>
+) {
+  const { label } = props;
+  if (label) {
+    return makeLabelRenderProps(data, props).map((labelRenderProps, index) => (
+      <Label key={index} renderLabel={label} labelProps={labelRenderProps} />
+    ));
+  }
 }
