+++ src/Label.tsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import type { CSSProperties, SVGProps } from 'react';
-import type { ExtendedData, BaseDataEntry } from './commonTypes';
+import type {
+  ExtendedData,
+  BaseDataEntry,
+  LabelRenderFunction,
+} from './commonTypes';
 
 export type LabelRenderProps<DataEntry extends BaseDataEntry> = {
   x: number;
@@ -13,11 +17,38 @@ export type LabelRenderProps<DataEntry extends BaseDataEntry> = {
   style?: CSSProperties;
 };
 
-type Props<DataEntry extends BaseDataEntry> = SVGProps<SVGTextElement> &
-  LabelRenderProps<DataEntry>;
+function ReactMinimalPieChartDefaultLabel<DataEntry extends BaseDataEntry>({
+  dataEntry,
+  dataIndex,
+  ...props
+}: SVGProps<SVGTextElement> & LabelRenderProps<DataEntry>) {
+  return <text dominantBaseline="central" {...props} />;
+}
 
 export default function ReactMinimalPieChartLabel<
   DataEntry extends BaseDataEntry
->({ dataEntry, dataIndex, ...props }: Props<DataEntry>) {
-  return <text dominantBaseline="central" {...props} />;
+>({
+  renderLabel,
+  labelProps,
+}: {
+  renderLabel: LabelRenderFunction<DataEntry>;
+  labelProps: LabelRenderProps<DataEntry>;
+}) {
+  const label = renderLabel(labelProps);
+  if (typeof label === 'string' || typeof label === 'number') {
+    return (
+      <ReactMinimalPieChartDefaultLabel
+        key={`label-${labelProps.dataEntry.key || labelProps.dataIndex}`}
+        {...labelProps}
+      >
+        {label}
+      </ReactMinimalPieChartDefaultLabel>
+    );
+  }
+
+  if (React.isValidElement(label)) {
+    return label;
+  }
+
+  return null;
 }
