+++ src/commonTypes.ts
@@ -1,11 +1,11 @@
-import type { ReactElement } from 'react';
+import type { ReactElement, ReactNode } from 'react';
 import type { LabelRenderProps } from './Label';
 
 export type EventHandler<Event> = (event: Event, dataIndex: number) => void;
 
 export type LabelRenderFunction<DataEntry extends BaseDataEntry> = (
   labelRenderProps: LabelRenderProps<DataEntry>
-) => number | string | ReactElement | undefined | null;
+) => ReactNode;
 
 export type BaseDataEntry = {
   title?: string | number;
